Move testing functions and musl-tests out
